Feature: Persona-backed evals.
We’ll replace the old “type” field in evals with a “feature” field, which specifies which feature we are testing. The idea is to utilize this field to run the evaluation through the Persona system.

As a first step, we are adding the option to list all existing features, which is sourced from DiscourseAi::Configuration::Module.